Modify minimal reporting
Removed the required output of jobid and state.  Allows the parsing of
single formats more easily and with format + should not be needed.
When a single format value is supplied the alignment is changed to left
justified and the header is not printed.  This should simplify piping to
another command (e.g. xargs grep)