fix: filter jobs with different states properly (#60)
* fix: filter jobs with different states properly

Closes #58 which noticed cases where state filtering would change
efficiency values.  The problem was that the output from sacct had a
state for the main job id (e.g. TIMEOUT) and a different one for
batch/extern (e.g. CANCELLED)

The fix is to filter states on only the main job ids then filter all
jobs by the main job id, instead of the state.

Fixed and added tests covering the change.

* chore: version bump